refactor(showcase): split app.vue in pages/ folder
